Create kolla_bridge.yml
This will help create bridges necessary.  It assumes that your eno1 is internal traffic (and left unmodified), and your eno2 (horizon interface and VM external traffic) is only configured for your controller nodes.  If your eno2 has an IP and is not a controller, then lines will be left intact and this will not deploy correctly.   
If deploying from MAAS, specify an IP for controllers.  For non-controller systems, leave the interface disconnected.